# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "An API Endpoint for handling organizational financial reimbursement requests",
        "title": "Financial Request Handler API",
        "contact": {
            "name": "APP Support",
            "email": "app_support@norainc.org"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "2.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/mileage": {
            "put": {
                "description": "allows a logged in user to edit their pending or reject mileage requests",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mileage",
                    "no-cache"
                ],
                "summary": "edit mileage request",
                "operationId": "edit-mileage",
                "parameters": [
                    {
                        "description": "mileage request information to update",
                        "name": "mileage-request-info",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.EditMileageInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.MileageOverviewRes"
                        }
                    }
                }
            },
            "post": {
                "description": "creates a mileage request for a logged in user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mileage",
                    "no-cache"
                ],
                "summary": "add mileage request",
                "operationId": "add-mileage",
                "parameters": [
                    {
                        "description": "new mileage request information",
                        "name": "mileage-request-info",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.MileageInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/responses.MileageOverviewRes"
                        }
                    }
                }
            },
            "delete": {
                "description": "allows a logged in user to delete and removes one of their mileage requests",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mileage",
                    "no-cache"
                ],
                "summary": "delete mileage request",
                "operationId": "delete-mileage",
                "parameters": [
                    {
                        "description": "mileage request id to delete",
                        "name": "mileage-request-id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.FindMileageInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.MileageOverviewRes"
                        }
                    }
                }
            }
        },
        "/mileage/approve": {
            "post": {
                "description": "allows an administrative user to approve a specific mileage request",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mileage",
                    "no-cache",
                    "admin"
                ],
                "summary": "approve mileage request",
                "operationId": "approve-mileage",
                "parameters": [
                    {
                        "description": "mileage request id to approve",
                        "name": "request-id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ApproveRejectRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.MileageOverviewRes"
                        }
                    }
                }
            }
        },
        "/mileage/detail": {
            "get": {
                "description": "generates detailed information for a specific mileage request",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mileage"
                ],
                "summary": "mileage request detail",
                "operationId": "mileage-detail",
                "parameters": [
                    {
                        "description": "mileage request id to find",
                        "name": "mileage-request-id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.FindMileageInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.MileageRes"
                        }
                    }
                }
            }
        },
        "/mileage/monthly": {
            "get": {
                "description": "creates a monthly mileage report for the entire organization for a given month and year",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mileage",
                    "reports",
                    "admin"
                ],
                "summary": "monthly mileage report",
                "operationId": "monthly-mileage",
                "parameters": [
                    {
                        "description": "month and year for report on organization wide mileage requests",
                        "name": "month-year-input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.MonthlyRequestInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.MileageOverviewsRes"
                        }
                    }
                }
            }
        },
        "/mileage/reject": {
            "post": {
                "description": "allows a logged in administrative user to reject a mileage request until further edits have been made",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mileage",
                    "no-cache",
                    "admin"
                ],
                "summary": "reject mileage request",
                "operationId": "reject-mileage",
                "parameters": [
                    {
                        "description": "mileage request id to reject",
                        "name": "request-id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ApproveRejectRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.MileageOverviewRes"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Action": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "user": {
                    "type": "string"
                }
            }
        },
        "models.ApproveRejectRequest": {
            "type": "object",
            "required": [
                "request_id"
            ],
            "properties": {
                "request_id": {
                    "type": "string"
                }
            }
        },
        "models.Category": {
            "type": "string",
            "enum": [
                "IOP",
                "INTAKE",
                "PEERS",
                "ACT_TEAM",
                "IHBT",
                "PERKINS",
                "MENS_HOUSE",
                "NEXT_STEP",
                "LORAIN",
                "PREVENTION",
                "ADMINISTRATIVE",
                "FINANCE"
            ],
            "x-enum-varnames": [
                "IOP",
                "INTAKE",
                "PEERS",
                "ACT_TEAM",
                "IHBT",
                "PERKINS",
                "MENS_HOUSE",
                "NEXT_STEP",
                "LORAIN",
                "PREVENTION",
                "ADMINISTRATIVE",
                "FINANCE"
            ]
        },
...
        "models.EditMileageInput": {
            "type": "object",
            "required": [
                "category",
                "date",
                "destination",
                "end_odometer",
                "grant_id",
                "id",
                "parking",
                "start_odometer",
                "starting_location",
                "tolls",
                "trip_purpose",
                "user_id"
            ],
            "properties": {
                "category": {
                    "$ref": "#/definitions/models.Category"
                },
                "date": {
                    "type": "string"
                },
                "destination": {
                    "type": "string"
                },
                "end_odometer": {
                    "type": "integer"
                },
                "grant_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "parking": {
                    "type": "number"
                },
                "start_odometer": {
                    "type": "integer"
                },
                "starting_location": {
                    "type": "string"
                },
                "tolls": {
                    "type": "number"
                },
                "trip_purpose": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "models.FindMileageInput": {
            "type": "object",
            "required": [
                "mileage_id"
            ],
            "properties": {
                "mileage_id": {
                    "type": "string"
                }
            }
        },
...
        "models.MileageInput": {
            "type": "object",
            "required": [
                "category",
                "date",
                "destination",
                "end_odometer",
                "grant_id",
                "parking",
                "start_odometer",
                "starting_location",
                "tolls",
                "trip_purpose"
            ],
            "properties": {
                "category": {
                    "$ref": "#/definitions/models.Category"
                },
                "date": {
                    "type": "string"
                },
                "destination": {
                    "type": "string"
                },
                "end_odometer": {
                    "type": "integer"
                },
                "grant_id": {
                    "type": "string"
                },
                "parking": {
                    "type": "number"
                },
                "start_odometer": {
                    "type": "integer"
                },
                "starting_location": {
                    "type": "string"
                },
                "tolls": {
                    "type": "number"
                },
                "trip_purpose": {
                    "type": "string"
                }
            }
        },
        "models.Mileage_Overview": {
            "type": "object",
            "properties": {
                "current_status": {
                    "type": "string"
                },
                "current_user": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "reimbursement": {
                    "type": "number"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "models.Mileage_Request": {
            "type": "object",
            "properties": {
                "action_history": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Action"
                    }
                },
                "category": {
                    "$ref": "#/definitions/models.Category"
                },
                "created_at": {
                    "type": "string"
                },
                "current_status": {
                    "type": "string"
                },
                "current_user": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "destination": {
                    "type": "string"
                },
                "end_odometer": {
                    "type": "integer"
                },
                "grant_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "last_user_before_reject": {
                    "type": "string"
                },
                "parking": {
                    "type": "number"
                },
                "reimbursement": {
                    "type": "number"
                },
                "start_odometer": {
                    "type": "integer"
                },
                "starting_location": {
                    "type": "string"
                },
                "tolls": {
                    "type": "number"
                },
                "trip_mileage": {
                    "type": "integer"
                },
                "trip_purpose": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "models.Month": {
            "type": "integer",
            "enum": [
                1,
                2,
                3,
                4,
                5,
                6,
                7,
                8,
                9,
                10,
                11,
                12
            ],
            "x-enum-varnames": [
                "January",
                "February",
                "March",
                "April",
                "May",
                "June",
                "July",
                "August",
                "September",
                "October",
                "November",
                "December"
            ]
        },
        "models.MonthlyRequestInput": {
            "type": "object",
            "required": [
                "month",
                "year"
            ],
            "properties": {
                "month": {
                    "$ref": "#/definitions/models.Month"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
...
        "responses.MileageOverviewRes": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/models.Mileage_Overview"
                },
                "status": {
                    "type": "string"
                },
                "string": {
                    "type": "string"
                }
            }
        },
        "responses.MileageOverviewsRes": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Mileage_Overview"
                    }
                },
                "status": {
                    "type": "string"
                },
                "string": {
                    "type": "string"
                }
            }
        },
        "responses.MileageRes": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/models.Mileage_Request"
                },
                "status": {
                    "type": "string"
                },
                "string": {
                    "type": "string"
                }
            }
        },
...
    }
}